refactor(stack-overflow-backend): remove alpha subpath
