+++ plugins/search-backend-module-stack-overflow-collator/README.md
@@ -87,9 +87,7 @@ import { createBackend } from '@backstage/backend-defaults';
 const backend = createBackend();
 backend.add(import('@backstage/plugin-search-backend/alpha'));
 backend.add(
-  import(
-    '@backstage/plugin-search-backend-module-stack-overflow-collator/alpha'
-  ),
+  import('@backstage/plugin-search-backend-module-stack-overflow-collator'),
 );
 backend.start();
 ```
