--- plugins/search-backend-module-stack-overflow-collator/alpha-api-report.md
@@ -1,11 +1,0 @@
-## API Report File for "@backstage/plugin-search-backend-module-stack-overflow-collator"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-import { BackendFeature } from '@backstage/backend-plugin-api';
-
-// @alpha
-const _default: () => BackendFeature;
-export default _default;
-```
