+++ plugins/search-backend-module-stack-overflow-collator/api-report.md
@@ -5,11 +5,16 @@
 ```ts
 /// <reference types="node" />
 
+import { BackendFeature } from '@backstage/backend-plugin-api';
 import { Config } from '@backstage/config';
 import { DocumentCollatorFactory } from '@backstage/plugin-search-common';
 import { IndexableDocument } from '@backstage/plugin-search-common';
 import { Logger } from 'winston';
 import { Readable } from 'stream';
+
+// @public
+const searchStackOverflowCollatorModule: () => BackendFeature;
+export default searchStackOverflowCollatorModule;
 
 // @public
 export interface StackOverflowDocument extends IndexableDocument {
