+++ plugins/search-backend-module-stack-overflow-collator/src/module/index.ts
@@ -14,10 +14,4 @@
  * limitations under the License.
  */
 
-/**
- * @packageDocumentation
- * A module for the search backend that exports Stack Overflow modules.
- */
-
-export * from './collators';
-export { searchStackOverflowCollatorModule as default } from './module';
+export { searchStackOverflowCollatorModule } from './SearchStackOverflowCollatorModule';
